Fixed vertical alignment of line numbers in search results on Mac
Due to different text positioning logic between the QItemDelegate and
the way SearchResultTreeItemDelegate drew the line numbers, the line
numbers did not align properly to the line of text next to it.

This was fixed by relying on QItemDelegate::drawDisplay for drawing the
line numbers, so that the same positioning logic is used.

This exposed a small right-alignment bug in QTextLayout, for which a fix
is coming to Qt.

Done-with: Erik Verbruggen
(cherry picked from commit 3bc20dbb60775e1e6a37b358f531ce83ec9cd2fd)