Fixed the default UI color
When there was no existing configuration, it would get set to pure black
on startup, resulting in a very dark gray Qt Creator. Now it is back to
its old friendly gray as default color.

Also made sure that the default on startup and the color used for
resetting the UI color in the settings are always in sync.

Reviewed-by: Jens Bache-Wiig