Open .prf files in the pro file editor
This fixes QTCREATORBUG-1434, but it's more of a workaround than a fix.
The problem was that the file was opened in the diff editor due to a
broken diff detection, and the diff editor in turn has a bug in that
it's kind of halfway read-only.

Reviewed-by: ossi
Task-number: QTCREATORBUG-1434
(cherry picked from commit 0afe76a67cc0987cd56e50903a5c5d7d60d3c873)