Wizard for standalone qml applications
Short: Create a new self running qml .sis/.deb package.

Long:
A typical Qml application usually consists of a couple of .qml,
.js and image files. Deploying such an application is challenging,
especially to mobile devices.

This wizard will generate a C++ project with a customized, light
weight Declarative Viewer, a helloworld Qml file and a .pro file
that takes care of the deployment of the Qml file(s).

This is just the first step. There will be an import wizard,
and further more there will be the possibility to add additional
Qml modules and probably also Qml directories to the package.

Task-number: QTCREATORBUG-1813