QmlJSOutline: Optimize performance of QmlOutlineModel
Previously we're setting the properties of the items one by one, each
time potentially triggering an update of the view. Actually this
uncovered an error in the Filter, which was only reacting to the first
change and therefore didn't take the item type into account on first
load.

Instead, we now use our own QStandardItem class which can be used to
assign all values at once.