QmlJS Rewriter: Allow disambiguation of where ObjectMember should be added to
With 700975ade3bf one could specify that a UiObjectMember should be inserted
after another one. But 0 'the default parameter' was ambiguous in that
either you want to have it inserted at the beginning of the member list,
or the heuristics should be applied.

Make this explicit by providing two methods.