QmlJS Live Preview: Showing object id's in context menu
To show the object string id's, the debug protocol is used to get them.
The problem is that the required methods in QDeclarativeContext(Data)
are not exposed, and it's too late to do it for 4.7.0. Hence this change
should be reverted in 4.8 when more efficient way of getting the id's
comes available.