Fix project-specific default file encoding
The project specific default file encoding would always be
locked to the default encoding when the project is loaded
the first time. This is unexpected and confuses users who
like to change the default encoding in the options dialog but
have never discovered (nor need) the project specific encoding.

The commit solves this by introducing a new value "Default" to
the project specific encoding combo box, which happens to
be the default.

Task-number: QTCREATORBUG-1996

Matthias