QmlProject: Search for qmlviewer in configured Qt versions (if not found in $PATH)
If no 'qmlviewer' executable is found in the PATH, iterate through the
list of configured Qt versions and try to find a qmlviewer there. This
should help users configuring creator such to play with Qml.

Right now the first qmlviewer found in a Qt version is selected. A UI
to let the user select one explicitly was not possible any more (string
freeze).

This requires a dependency from QmlProjectManager to
Qt4ProjectManager.

Reviewed-by: dt