Get rid of Qt4ProFileNode in LibraryDetailsController
Since it will not work for pri files. The code is prepared to work
with pri files. However, currently it will work only in case
the pro file which includes the pri file lays in the same dir.
After the QTBUG-13057 is done it will be just a matter of
changing the apropriate variable name in librarydetailscontroller.cpp
to make it working properly.

Task-number: QTCREATORBUG-125