Maemo: Delay UTFS server start a bit.
We have no way to know when the client is actually opening its ports,
so we wait until we are reasonably sure it's done.
This is not 100% safe, but should generally suffice.
(The UTFS server seems to have this logic already built in, but it does
not seem to work on Mac.)