Remove one hack of the QTipLabel hack.
This code used two hacks: it created a class with
metaObject()->className() == "QTipLabel" to fool the code in Qt into
applying the style properly to this class.

But it did that by creating a class called QTipLabel. This works as
long as the class in Qt is hidden by symbol visibility. When symbol
visibility is disabled (for whatever reason), then Creator simply
crashes due to size mismatches and memory corruptions (virtual tables,
etc.)

This patch makes the class be called TextEditor::Internal::QTipLabel,
but fools moc into thinking it isn't.

A proper solution would be to simply override the meta object.

Signed-off-by: Leandro Melo <leandro.melo@nokia.com>