Remove EditorManager::ensureEditorManagerVisible()
The function always switched to Edit Mode, which was the cause for
several bugs. Since openEditor() already does switch to the right mode,
the function is not really needed.

The following actions now respect the prefered mode of the editor:
"f file"-locator
The Open Documents list
Editor history navigation