Do not force a specific editor for a file created from a wizard.
The wizards already provide a default suffix for file creation.
But since the user is still free to specify something else it's better to leave it to Creator to decide which one is the proper editor.
After all, this reflects how the file will be treated by Creator (including the next time it's opened).

Task-number: QTCREATORBUG-2006