Qt4ProjectManager: Remove run dependency on QmlJSEditor
The Qt4ProjectManager still depends on the QMLJsEditor for building.
Since the Qt4ProjectManager does cope correctly with not
finding a ModelManager, the Qt4ProjectManager will continue to work even
without the qmljseditor.

Note: The load order between qt4projectmanager and qmljseditor is undefined
now. This is not a problem for this particular patch, but might break
future code in subtle ways.