debugger: fix QObject property dumping on Windows
The problem of gdb finding QVector::malloc instead of the global malloc
appeared again. The workaorund is to use calloc which is not used by Qt.
However, for transfering string literal data to the inferior gdb uses
malloc internally outside our control, so we cannot use string literals
in inferior calls at all.

In the particular case of QObject property names we are lucky as the
data already exists in the inferior memory. This patch just uses it.

Reviewed-by: Leandro Melo

Task: QTCREATORBUG-2533