Merged 'New Qt quick App' and 'Import Qml App' Wizards
The only difference between these wizards was that in the
Import Wizard, the user was able to chose a .qml file.
So, why not merge them. Also, this commit is a nice code
reduction.

The module deployment feature was killed by popular demand.
Apparently, nobody writes or deploys Qml modules, these days.
Let's see when they want to have it back ;)

Task-Number: QTCREATORBUG-2505
Reviewed-by: ck