No Samba while including files
(The other Samba is meant. The slow one, not the nice one.)

Creator manages to feed the CppModelManager with all kinds of
wrong include paths. That is usually not a big issue as long
as that path is not a simple "/".

CppManager appended a separator to all incoming include paths.
And when the above "/" was passed in, we had (on my machine)
7 seconds of Samba Network lookup, multiple times during a
parse. And the sad thing was that my dear colleagues in the
Intranet did not even have the searched-for header files on
their shared folders.

This patch makes sure that a separator is only appended if
the path does not already end with one.

The real core of the issue (wrong include paths from Creator)
still needs to be fixed.

Done-with: Roberto Raggi