Text editor: Fix potential crash in the plain text editor.
Example situation: The plain text editor is displaying in-memory content and
has not been explicitly "configured". If there's an attempt to indent the text
it crashes since the indenter is invalid. This can happen if one clicks on 'Show
View Contents in Editor' within a debug session and press enter in the editor.

Reviewed-by: con