Prevent crash when closing Qt Creator while cdb session is running.
The debugger plugin opens an editor in this case, leading to fakevim and
cppeditor crashing.
This commit doesn't fix the problem's source, but adds safeguards that
are sensible anyhow: FakeVim should not register with editors during
shutdown, and CppEditor should not unconditionally assume an existing
editor manager in the destructor.

Reviewed-by: Friedemann Kleint