QmlProject: Don't use RegExp for simple wildcard matching
QRegExp matching is quite expensive, and has to be done for every file
in the project directory tree against all possible suffixes. Optimize
for the common case that the pattern is "*.suffix" by doing a
fileName.endsWidth(suffix) in this case.

This speeds up loading of examples/declarative/declarative.qmlproject by
about 30%.

Reviewed-by: Christian Kamm