Observer mode: Fix automatic selection of all instances
If the user clicks on an instance of an element in the running app
while in server mode, the cursor selection is automatically updated.
However, the change of the cursor then leads to all the other
instances of the element at the cursor position being selected too.

Prevent this 'call back' in the QmlInspector by checking whether one
of the newly selected items is the one just selected.

Simpler mechanisms (like a sequential blocking of updates) don't
work because the call back happens through a QTimer.

Task-number: QTCREATORBUG-2366