Generic highlighter: Add Creator's share folder in the fallback definitions search.
This leaves the code prepared if we decide to ship some of the Kate XMLs (still
a pending decision). In this case the installer needs to be "notified" and the
Kate files put in the repo.

An addition check for actual XML files in the searched directories is also added.

Reviewed-by: con
Reviewed-by: Thorbjorn Lindeijer