QmlObserver: Remove breadcrumb widget in observer mode
I wasn't able to get it working on Mac OS X: Putting it on top
of the QGraphicsView doesn't work because Cocoa doesn't allow a
widget to be on top of OpenGL viewport. Embedding it as a
QGraphicsProxyWidget produces lots of Cocoa warnings (raster engine related).

Anyhow, the same widget shows up in the Debugger/Observer tab of
Qt Creator.

Reviewed-by: Lasse Holmstedt