QmlObserver: Fix navigation on Mac OS X
The QmlObserver relies on installing an event handler on the graphicsview
viewport, to get e.g. Mouse events.

QmlObserver on Mac OS X switches the viewport later on to a QGLViewPort, when
QmlObserver has already been set up ... Make QmlObserver aware of this by
listening to the ChildRemoved event that will be signalled in this case.

Task-number: QTCREATORBUG-2892
Reviewed-by: Lasse Holmstedt