Remove empty dir from static.pro
This was added in the case we start shipping some of the Kate XMLs (still
a pending decision), which should be placed in this dir. However, this
causes a warning when loading qtcreator.pro with master branch version.
So I'll removed it. If we really ship the files it should be re-added
and also the installer notified.

Reviewed-by: con