Do not link against (the possibly absent) eiksrv.dso
The template .pri file of the "Qt Quick Application" and
"Qt Mobile Application" wizards linked against eiksrv
when locking the screen orientation. That is however not
necessary.

The real problem is that eiksrv does not get shipped with
the Symbian^1 SDK in Nokia Qt SDK 1.0/1.1. Therefore, the
innecessary link attempt even fails.

This patch removes the -leiksrv and bumps the version for
the template.

Cherrypicking into qtcreator/2.1 is highly recommended.

Task-Number: QTCREATORBUG-3142
Task-Number: QTSDK-194