Fix command line parsing: Do return empty items.
Symbians Makefiles have a "# Command:" line in the header
which has two whitespaces in the middle. The line parsing bug
fixed by this commit lead to a discrepency of the
"Actual args" (without empty elements) and the "Parsed args"
(with empty elements, otherwise same to Actual args), and
thus to a call of qmake on each build.

Task-Number: QTBUG-15539