QmlDesigner.rewriter: fix possible incompatibilities with import plugins
The current Qt Version might be build with mingw or even be a Symbian
or Maemo target. This patch ensures that we do not try to load a plugin
from the configured Qt Version anymore.
Instead we only load plugins provided by Qt Creator
or plugins which are specified by:  QML_IMPORT_PATH

Reviewed-by: Kai Koehne
Reviewed-by: Erik Verbruggen