Handle read only files correctly when saving all modified files
We use the editor manager's way to save modified files now, so the
normal logic for read only files is applied, including the "cancel/save
as/open vcs/make writable" dialog.
Also fixes that it wasn't handled as a "cancel" if the user chooses to
cancel this dialog.

Task-number: QTCREATORBUG-3109