Add "never ask again" option to dialog asking about killing applications
The option is used for the stop button in the application output.
Closing tabs will still ask unconditionally, also if applications would
be killed as a side-effect, e.g. of closing Qt Creator.
The Symbian controls also always ask, with a more descriptive text,
because killing applications might have negative effect on the device in
some cases.

Reviewed-by: Friedemann Kleint
Task-number: QTCREATORBUG-3119