Dummy template for Html5 app wizards
The template code is similar to the Qt Quick App wizard
template code. It has a Viewer class +.pri which wrap the "dirty
stuff", a main .pro file and main.cpp file. In this case, there
is also a "Hello World" index.html file.

The "dirty stuff" code exposes a quit() function to the JavaScript
in the webview, which can be triggerd via the index.html.

As the 1-line-summary of this commit states, it is a dummy. Some-
one with proper knowlede about QtWebkit please provide better
template code.

Task-Number: QTCREATORBUG-3284