QmlDesigner.NodeInstances: Add dummydata per file
A component which is asking for the e.g. parent.width in root object
has mostly not the wanted size. To get around this problem you can put
a file in the dummydata directory with the signature:

originalfilename_dummydata.qml:


import QtQuick 1.0
import QmlDesigner 1.0

DummyContextObject {
    parent: QtObject {
        property real width: 1000
        property real height: 400
    }
}

The file is reloaded if you change it so you can make changes on the fly.