don't give up on evaluation errors too easily
if the exact evalution fails (possibly because of a parser error), then
still use the cumulative evaluation to construct the project tree. only
the target information won't be updated and the code model will be left
in the dark.

todo: the errors are dumped only to the (unnumbered) "general messages"
output pane, which makes them practically invisible.

Task-number: QTCREATORBUG-3093