Provide way to show a variable chooser widget.
For the Qt Creator variables, currently used in the external tools.
The variable chooser looks if the current focus widget has a variable
support property set and is a line edit, text edit or plain text edit.
For line edits it adds a little icon button that shows the chooser.