F1+F1 behavior was broken if there was no context help found.
If on F1 no documentation was found, the focus was not set to the side
bar, making it impossible to press F1 again to get into help mode.
The fix also makes sure that the "blank" page is not set to the help
mode in this case, to avoid a big white empty area.