Editors: Change indent behavior for single line selection
When only one line is selected when press the TAB key, should delete
the selected contents, then perform indent.
It like other more editor behavior.

Reviewer notes: Indeed this seems to be a common behavior in other
editors (for example Visual Studio, Code Blocks, Eclipse). But in
order to preserve the indentation consistent with the multiple-line
case it would be nice an approach like Eclipse's: If the single
line is completely selected, do the indentation as usual. Otherwise
remove the selected text.

The author of the request is not able to revise this minor change
because he claims to be busy (and actually asked for the help). So
I will keep this patch in the original state and implement the
detail I mentioned above in a following commit.

Merge-request: 252
Reviewed-by: Leandro Melo <leandro.melo@nokia.com>