Observer: Fixed synchronization of animation speed and paused state
The menus of the Observer were not being updated correctly when
animation speed was changed or when the animations were paused.

Now the whole thing looks a bit cleaner and the Observer menu
uses the QDeclarativeViewObserver to change these settings and stays in
sync when they change from other places.

Reviewed-by: Kai Koehne