QmlDesigner: fixes state editing in text
In the stateseditor we have to use propertiesRemoved instead of
propertiesAboutToBeRemoved when reseting the model. Otherwise
the change we want to react to has not happened, yet.
Some sanity checking in the property editor.

In the DesignDocumentController we slightly change the order.
The textMoifider is activated (again) before we set back the old
state.
Note: We do check for isValid() since the rewriter might delete the state.


Reviewed-by: Kai Koehne
Task-number: QTCREATORBUG-4199