Fix progress bar for searching in files.
Affects e.g. Advanced Find->All Projects. Progress bar was not updating
at all, except at the end of the search.
The fix is a workaround for regression in
QFutureInterface::isProgressUpdateNeeded that would no longer return
true the very first time in Qt 4.7. To be reverted when this is fixed
in a released Qt version.

Reviewed-by: dt