Fix deployment with MinGW and sh.exe in path
When building with MinGW and sh.exe is in the path, we should/can
not use xcopy in our deployment code. This patch detects the
presence of sh.exe and adjusts the behavior accordingly.

'cp -r' is used directly instead of $(COPY_DIR), because the fix
for QTBUG-17315 is not yet in a released Qt version.

'mkdir -p' is used directly instead of $(MKDIR), because the
QMAKE_MKDIR for win32-g++ (in case of QMAKE_SH) does not have
the '-p'. A least not til Qt 4.7.3.

Task-Number: QTCREATORBUG-3094
Reviewed-by: Oswald Buddenhagen