QmlDesigner.copyNPaste: fix for Components and default properties
To handle defaultProperties correctly in the text buffer for
copy n paste we have to forward the meta system.
Since the text buffer is not a qml document in Qt Creator
with all the infrastructure attached the meta system is
incomplete.
So we forward in these cases the meta system of the actual document.

Model::create() now gets an additional paramater where we can specify another
model for proxying the meta system.

Reviewed-by: Kai Koehne