reshuffle word counting in expressions
this makes the value list sizehints more correct: instead of the number
of literals and expansions, it's now the number of actual words, and
that only if the expression starts with a literal (otherwise the
pre-allocation is discarded anyway).
the performance impact seems negligible, with qt being a tad slower and
creator a tad faster.

as a side effect, complaining about excess words in a test expression
does not immediately terminate the current line's parsing.