continue parsing line despite errors
as far as possible, anyway. and suppress messages after the first one.
this avoids that scope nesting gets thrown off and thus spurious
followup-errors crop up.

as a "side effect", this fixes a crash resulting from us rewinding the
token stream but ignoring the block stack on the way.

Task-number: QTCREATORBUG-4368