Another fix for shadow build deployment with sh.exe in path
The "isEmpty(QMAKE_SH)" test is nonsense in the case of an msvc
build, because for msvc, sh.exe is never executed (although QMAKE_SH
is set.

"isEqual(QMAKE_DIR_SEP, \\)" is now the ultimate test. QMAKE_DIR_SEP
will only be "/" if sh.exe is in the path AND win32-g++ is used.

Task-Number: QTCREATORBUG-4455