Debugger: Have Python dumper print type size information.
For memory views. In the dumper, create a cache of TypeInfo
structs to contain type information and print the new entries
in each call of bb().

Collect the information in the gdb engine.
Replace WatchData::origAddress by WatchData::ReferencingAddress
to be able to correctly handle automatically dereferenced
pointers of the gdb engine. Whereas 'address' of a automatically
dereferenced used to be that of the pointer, it is now that
of the dereferenced item matching the size reported, enabling
memory views. Referencing address is now the address of the pointer.

Reviewed-by: hjk