Prevent improper sizing of MiniProjectTargetSelector
The selector would sometimes be too high (if projects were removed),
and sometimes show vertical scrollbars (if projects were added). This
problem was limited to the first time that the widget was opened on
all platforms but on OS X, where it could be observed persistently
and where its lack potentially can cause problems due to ed5d10a.

Reviewed-by: con
Task-Number: QTCREATORBUG-1792