Fix qt version number in Qt Quick Wizard
Temporal lapse in understanding which qt version is required for the Qt
Quick Wizard. On Maemo we have the Qt Quick compat plugin which allows
the generated project to work with 4.7.0. In theory that doesn't apply
to symbian, but qt 4.7.0 does not support symbian. (And on the desktop
everyone has upgraded.)

Reviewed-By: ck