QmlProject: Fix import of .qmlproject.user files from 2.1
We changed the semantics of an empty 'mainScript' in the .user file:
In 2.1 it translated to 'use current file', while in 2.2 the default is now
to use the file specified in .qmlproject.

However, if there is no mainFile specified in .qmlproject we should
fall back to the current file in editor.

Reviewed-by: Christiaan Janssen