QmlDesigner.qmlPuppet: fixes a bug when dragging in a WebView on KDE
WebView object contains a QIcon property ("reload.icon"),
which fails to be streamed from the puppet (client) to Qt Creator (server).
The QIcon is created by a 'KIconEngine' in the KDE case.

which isn't found on the QtCreator side.
This leads operator>> in qicon.cpp to not read anything from
the buffer at all ... leading to all following properties being
invalid, since read from the wrong bytes in the stream.

Solution: Since QIcon leads to problems and we have no current use case
for a QIcon property in Bauhaus we just ignore properties of type QIcon.

Reviewed-by: Marco Bubke
Task-number: QTCREATORBUG-4596