Maemo: Use different desktop files for Fremantle and Harmattan.
So far, we desperately tried not to duplicate the desktop file,
resorting to all kinds of "clever" heuristics in the process.
This is, however, somewhat limiting and confusing to users,
because we do magic stuff behind their backs.
So it seems better to "give up" and use two desktop files.

Change-Id: Idf969eb3ad365f080e5a9c1608149ea6de629d5c
Reviewed-on: http://codereview.qt.nokia.com/31
Reviewed-by: Qt Sanity Bot <qt_sanity_bot@ovi.com>
Reviewed-by: Alessandro Portale <alessandro.portale@nokia.com>