Completion: Ignore non identifiers when matching proposals
Since now we reduced the idle editor time for showing completions,
some things started to be a bit annoying. This fix changes the
prefix from the proposal to include only letters, digits, and the
underscore. Yes, technically they are not allways identifiers
in a generic sense, but it should be enough for our purpose.

Now, you should not receive a proposal 'foo:' when you have already
typed 'foo' in QML, for example.

Change-Id: Ica92182a34636598faedb067d0527e37ca6fee89
Reviewed-on: http://codereview.qt.nokia.com/46
Reviewed-by: Roberto Raggi <roberto.raggi@nokia.com>